test: consolidate tag inheritance tests into test_tag_inheritance.py
Move all tag-inheritance-related tests into a single file:
- InheritedTagsTests, InheritedTagsImportMixin, InheritedTagsImportTestAPI,
  InheritedTagsImportTestUI from test_tags.py
- TestTagInheritanceOnPersist from test_bulk_locations.py
- test_tags_signals.py (new unit tests, now absorbed)

Add new unit test classes covering tags_signals.py internals:
- TestGetProducts — isinstance routing for all model types
- TestInheritProductTags — per-product and system-wide flag gating
- TestPropagateInheritanceEarlyExit — early-exit short-circuit logic

Add class-level docstrings throughout for clarity.
Fix SSL redirect in InheritedTagsImportTestAPI/UI setUp.
