+++ unittests/test_tags.py
@@ -454,239 +454,3 @@ class TagImportTestUI(DojoAPITestCase, TagImportMixin):
             self.assertEqual(302, response.status_code, response.content[:1000])
             new_test_id = int(response.url.split("/")[-1])
             return {"test": new_test_id}
-
-
-@override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-@versioned_fixtures
-class InheritedTagsTests(DojoAPITestCase):
-
-    """Non-import tests for inherited tags functionality."""
-
-    fixtures = ["dojo_testdata.json"]
-
-    def setUp(self, *args, **kwargs):
-        super().setUp()
-        self.login_as_admin()
-        self.system_settings(enable_product_tag_inheritance=True)
-        self.product = self.create_product("Inherited Tags Test", tags=["inherit", "these", "tags"])
-        self.scans_path = get_unit_tests_scans_path("zap")
-        self.zap_sample5_filename = self.scans_path / "5_zap_sample_one.xml"
-
-    def _convert_instance_tags_to_list(self, instance) -> list:
-        return [tag.name for tag in instance.tags.all()]
-
-    def test_new_engagement_then_add_tag_to_engagement_then_remove_tag_to_engagement(self):
-        # Create the engagement
-        engagement = self.create_engagement("Inherited Tags Engagement", self.product)
-        test = self.create_test(engagement=engagement, scan_type="ZAP Scan")
-        # Check to see if tags match the product
-        product_tags = self._convert_instance_tags_to_list(self.product)
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(engagement))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(test))
-        # Add a tag on the engagement)
-        engagement_tags_before_addition = self._convert_instance_tags_to_list(engagement)
-        engagement.tags.add("engagement_only_tag")
-        # Check to see that the update was successful
-        self.assertEqual(["engagement_only_tag", *engagement_tags_before_addition], self._convert_instance_tags_to_list(engagement))
-        # Check to see that tests were not impacted
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(test))
-        # remove a tag on the engagement
-        engagement_tags_before_removal = self._convert_instance_tags_to_list(engagement)
-        engagement.tags.remove("engagement_only_tag")
-        # Check to see that the update was successful
-        engagement_tags_before_removal.remove("engagement_only_tag")
-        self.assertEqual(engagement_tags_before_removal, self._convert_instance_tags_to_list(engagement))
-        # Check to see that tests were not impacted
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(test))
-
-    def test_new_engagement_then_remove_inherited_tag(self):
-        # Create the engagement
-        engagement = self.create_engagement("Inherited Tags Engagement", self.product)
-        # Check to see if tags match the product
-        product_tags = self._convert_instance_tags_to_list(self.product)
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(engagement))
-        # Remove an inherited tag
-        engagement_tags_before_removal = self._convert_instance_tags_to_list(engagement)
-        engagement.tags.remove("inherit")
-        # Check to see that the inherited tag could not be removed
-        self.assertEqual(engagement_tags_before_removal, self._convert_instance_tags_to_list(engagement))
-
-
-class InheritedTagsImportMixin:
-
-    """Mixin containing inherited tags import/reimport tests that can be run via API or UI."""
-
-    def setUp(self):
-        self.system_settings(enable_product_tag_inheritance=True)
-        self.product = self.create_product("Inherited Tags Test", tags=["inherit", "these", "tags"])
-        self.scans_path = get_unit_tests_scans_path("zap")
-        self.zap_sample5_filename = self.scans_path / "5_zap_sample_one.xml"
-
-    def _convert_instance_tags_to_list(self, instance) -> list:
-        return [tag.name for tag in instance.tags.all()]
-
-    def _import_and_return_objects(self, test_id=None, *, reimport=False, tags=None) -> dict:
-        # Import some findings to create all objects
-        engagement = self.create_engagement("Inherited Tags Engagement", self.product)
-        if reimport:
-            response = self.reimport_scan_with_params(test_id, self.zap_sample5_filename, tags=tags)
-        else:
-            response = self.import_scan_with_params(self.zap_sample5_filename, engagement=engagement.id, tags=tags)
-
-        test_id = response["test"]
-        test = Test.objects.get(id=test_id)
-        finding = Finding.objects.filter(test=test).first()
-        location = self._get_location(finding)
-        return {
-            "engagement": engagement,
-            "location": location,
-            "test": test,
-            "finding": finding,
-        }
-
-    def _get_location(self, finding):
-        # TODO: Delete this after the move to Locations
-        if not settings.V3_FEATURE_LOCATIONS:
-            return finding.endpoints.all().first()
-        return finding.locations.all().first().location
-
-    def test_import_without_tags(self):
-        # Import some findings to create all objects
-        objects = self._import_and_return_objects()
-        # Check that the tags all match what the product has
-        product_tags = self._convert_instance_tags_to_list(self.product)
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("engagement")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("location")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("test")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("finding")))
-
-    def test_import_with_tags_then_reimport_with_different_tag(self):
-        # Import some findings to create all objects
-        objects = self._import_and_return_objects(tags=["import_tag"])
-        # Check that the tags all match what the product has
-        product_tags = self._convert_instance_tags_to_list(self.product)
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("engagement")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("location")))
-        self.assertEqual(["import_tag", *product_tags], self._convert_instance_tags_to_list(objects.get("test")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("finding")))
-        # Reimport now
-        objects = self._import_and_return_objects(test_id=objects.get("test").id, reimport=True, tags=["reimport_tag"])
-        # Check that the tags all match what the product has
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("engagement")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("location")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("finding")))
-        # Make a copy of the list becase of the need for the lists to be exact (index for index)
-        product_tags_plus_reimport_tag = product_tags.copy()
-        product_tags_plus_reimport_tag.insert(1, "reimport_tag")
-        self.assertEqual(product_tags_plus_reimport_tag, self._convert_instance_tags_to_list(objects.get("test")))
-
-    def test_remove_tag_from_product_then_add_tag_to_product(self):
-        # Import some findings to create all objects
-        objects = self._import_and_return_objects()
-        # Check that the tags all match what the product has
-        product_tags = self._convert_instance_tags_to_list(self.product)
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("engagement")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("location")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("test")))
-        self.assertEqual(product_tags, self._convert_instance_tags_to_list(objects.get("finding")))
-        # Remove a tag from the product
-        self.product.tags.remove("inherit")
-        # This triggers an async function with celery that will fail, so run it manually here
-        propagate_tags_on_product_sync(self.product)
-        # Save the tags post removal
-        product_tags_post_removal = self._convert_instance_tags_to_list(self.product)
-        # Check that the tags all match what the product has
-        self.assertEqual(product_tags_post_removal, self._convert_instance_tags_to_list(objects.get("engagement")))
-        self.assertEqual(product_tags_post_removal, self._convert_instance_tags_to_list(objects.get("location")))
-        self.assertEqual(product_tags_post_removal, self._convert_instance_tags_to_list(objects.get("test")))
-        self.assertEqual(product_tags_post_removal, self._convert_instance_tags_to_list(objects.get("finding")))
-        # Add a tag from the product
-        self.product.tags.add("more", "tags!")
-        # This triggers an async function with celery that will fail, so run it manually here
-        propagate_tags_on_product_sync(self.product)
-        # Save the tags post removal
-        product_tags_post_addition = self._convert_instance_tags_to_list(self.product)
-        # Check that the tags all match what the product has
-        self.assertEqual(product_tags_post_addition, self._convert_instance_tags_to_list(objects.get("engagement")))
-        self.assertEqual(product_tags_post_addition, self._convert_instance_tags_to_list(objects.get("location")))
-        self.assertEqual(product_tags_post_addition, self._convert_instance_tags_to_list(objects.get("test")))
-        self.assertEqual(product_tags_post_addition, self._convert_instance_tags_to_list(objects.get("finding")))
-
-
-@override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-@versioned_fixtures
-class InheritedTagsImportTestAPI(DojoAPITestCase, InheritedTagsImportMixin):
-
-    """Test inherited tags during import/reimport via API."""
-
-    fixtures = ["dojo_testdata.json"]
-
-    def setUp(self):
-        super().setUp()
-        testuser = User.objects.get(username="admin")
-        testuser.usercontactinfo.block_execution = True
-        testuser.usercontactinfo.save()
-        self.login_as_admin()
-        InheritedTagsImportMixin.setUp(self)
-
-
-@override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-@versioned_fixtures
-class InheritedTagsImportTestUI(DojoAPITestCase, InheritedTagsImportMixin):
-
-    """Test inherited tags during import/reimport via UI."""
-
-    fixtures = ["dojo_testdata.json"]
-
-    def setUp(self):
-        super().setUp()
-        testuser = User.objects.get(username="admin")
-        testuser.usercontactinfo.block_execution = True
-        testuser.usercontactinfo.save()
-        self.login_as_admin()
-        self.client_ui = Client()
-        self.client_ui.force_login(self.get_test_admin())
-        InheritedTagsImportMixin.setUp(self)
-
-    def import_scan_with_params(self, filename, scan_type="ZAP Scan", engagement=1, minimum_severity="Low", *,
-                                active=True, verified=False, tags=None, close_old_findings=False, **kwargs):
-        """Override to use UI import instead of API."""
-        with Path(filename).open(encoding="utf-8") as testfile:
-            payload = {
-                "minimum_severity": minimum_severity,
-                "active": "force_to_true" if active else "force_to_false",
-                "verified": "force_to_true" if verified else "force_to_false",
-                "scan_type": scan_type,
-                "file": testfile,
-                "environment": 1,
-                "close_old_findings": close_old_findings,
-            }
-            if tags is not None:
-                # Tagulous form field expects comma-separated string
-                payload["tags"] = ",".join(tags) if isinstance(tags, list) else tags
-
-            response = self.client_ui.post(reverse("import_scan_results", args=(engagement,)), payload)
-            self.assertEqual(302, response.status_code, response.content[:1000])
-            test_id = int(response.url.split("/")[-1])
-            return {"test": test_id}
-
-    def reimport_scan_with_params(self, test_id, filename, scan_type="ZAP Scan", minimum_severity="Low", *,
-                                  active=True, verified=False, tags=None, close_old_findings=True, **kwargs):
-        """Override to use UI reimport instead of API."""
-        with Path(filename).open(encoding="utf-8") as testfile:
-            payload = {
-                "minimum_severity": minimum_severity,
-                "active": "force_to_true" if active else "force_to_false",
-                "verified": "force_to_true" if verified else "force_to_false",
-                "scan_type": scan_type,
-                "file": testfile,
-                "close_old_findings": close_old_findings,
-            }
-            if tags is not None:
-                # Tagulous form field expects comma-separated string
-                payload["tags"] = ",".join(tags) if isinstance(tags, list) else tags
-
-            response = self.client_ui.post(reverse("re_import_scan_results", args=(test_id,)), payload)
-            self.assertEqual(302, response.status_code, response.content[:1000])
-            new_test_id = int(response.url.split("/")[-1])
-            return {"test": new_test_id}
